// migrations/snapshots/e631114033910891e449f2bb033bdde2c659fb4fac022701949e6c0c79de6dd1/contract.d.ts
export type CodecTypes = PgTypes;
export type LaneCodecTypes = CodecTypes;
export type QueryOperationTypes = PgAdapterQueryOps<CodecTypes>;
// ...
export type FieldOutputTypes = {
  readonly public: {
    readonly Approval: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly eventId: CodecTypes['pg/int4@1']['output'];
      readonly approverId: CodecTypes['pg/int4@1']['output'];
      readonly status: 'PENDING' | 'APPROVED' | 'REJECTED';
      readonly reason: CodecTypes['pg/text@1']['output'] | null;
    };
    readonly Booking: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly eventId: CodecTypes['pg/int4@1']['output'];
      readonly venueId: CodecTypes['pg/int4@1']['output'];
      readonly amount: CodecTypes['pg/int4@1']['output'];
      readonly taxAmount: CodecTypes['pg/int4@1']['output'];
      readonly currency: CodecTypes['pg/text@1']['output'];
      readonly paymentStatus: 'PENDING' | 'PAID' | 'FAILED' | 'REFUNDED';
      readonly razorpayOrderId: CodecTypes['pg/text@1']['output'] | null;
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['output'];
    };
    readonly Company: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
      readonly slug: CodecTypes['pg/text@1']['output'];
    };
    readonly Event: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly companyId: CodecTypes['pg/int4@1']['output'];
      readonly createdById: CodecTypes['pg/int4@1']['output'];
      readonly title: CodecTypes['pg/text@1']['output'];
      readonly eventType: CodecTypes['pg/text@1']['output'];
      readonly purpose: CodecTypes['pg/text@1']['output'] | null;
      readonly city: CodecTypes['pg/text@1']['output'];
      readonly dateTime: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly attendees: CodecTypes['pg/int4@1']['output'];
      readonly budget: CodecTypes['pg/int4@1']['output'];
      readonly diet: CodecTypes['pg/text@1']['output'];
      readonly status:
        | 'DRAFT'
        | 'REQUESTED'
        | 'APPROVED'
        | 'VENUE_SELECTED'
        | 'REJECTED'
        | 'BOOKED'
        | 'COMPLETED'
        | 'CANCELLED';
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['output'];
    };
    readonly Invoice: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly bookingId: CodecTypes['pg/int4@1']['output'];
      readonly pdfUrl: CodecTypes['pg/text@1']['output'];
      readonly gstin: CodecTypes['pg/text@1']['output'];
    };
    readonly Policy: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly companyId: CodecTypes['pg/int4@1']['output'];
      readonly maxBudget: CodecTypes['pg/int4@1']['output'];
      readonly perPersonCap: CodecTypes['pg/int4@1']['output'];
      readonly allowedCities: ReadonlyArray<CodecTypes['pg/text@1']['output']>;
      readonly approverRole: 'ADMIN' | 'REQUESTER' | 'APPROVER' | 'FINANCE';
    };
    readonly User: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly clerkId: CodecTypes['pg/text@1']['output'];
      readonly email: CodecTypes['pg/text@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
      readonly role: 'ADMIN' | 'REQUESTER' | 'APPROVER' | 'FINANCE';
      readonly companyId: CodecTypes['pg/int4@1']['output'] | null;
    };
    readonly Venue: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly companyId: CodecTypes['pg/int4@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
      readonly city: CodecTypes['pg/text@1']['output'];
      readonly address: CodecTypes['pg/text@1']['output'] | null;
      readonly capacity: CodecTypes['pg/int4@1']['output'];
      readonly cuisine: CodecTypes['pg/text@1']['output'];
      readonly priceBand: CodecTypes['pg/text@1']['output'];
      readonly tags: ReadonlyArray<CodecTypes['pg/text@1']['output']>;
      readonly rating: CodecTypes['pg/float8@1']['output'];
      readonly active: CodecTypes['pg/bool@1']['output'];
    };
  };
};
export type FieldInputTypes = {
  readonly public: {
    readonly Approval: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly eventId: CodecTypes['pg/int4@1']['input'];
      readonly approverId: CodecTypes['pg/int4@1']['input'];
      readonly status: 'PENDING' | 'APPROVED' | 'REJECTED';
      readonly reason: CodecTypes['pg/text@1']['input'] | null;
    };
    readonly Booking: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly eventId: CodecTypes['pg/int4@1']['input'];
      readonly venueId: CodecTypes['pg/int4@1']['input'];
      readonly amount: CodecTypes['pg/int4@1']['input'];
      readonly taxAmount: CodecTypes['pg/int4@1']['input'];
      readonly currency: CodecTypes['pg/text@1']['input'];
      readonly paymentStatus: 'PENDING' | 'PAID' | 'FAILED' | 'REFUNDED';
      readonly razorpayOrderId: CodecTypes['pg/text@1']['input'] | null;
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['input'];
    };
    readonly Company: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
      readonly slug: CodecTypes['pg/text@1']['input'];
    };
    readonly Event: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly companyId: CodecTypes['pg/int4@1']['input'];
      readonly createdById: CodecTypes['pg/int4@1']['input'];
      readonly title: CodecTypes['pg/text@1']['input'];
      readonly eventType: CodecTypes['pg/text@1']['input'];
      readonly purpose: CodecTypes['pg/text@1']['input'] | null;
      readonly city: CodecTypes['pg/text@1']['input'];
      readonly dateTime: CodecTypes['pg/timestamptz-string@1']['input'];
      readonly attendees: CodecTypes['pg/int4@1']['input'];
      readonly budget: CodecTypes['pg/int4@1']['input'];
      readonly diet: CodecTypes['pg/text@1']['input'];
      readonly status:
        | 'DRAFT'
        | 'REQUESTED'
        | 'APPROVED'
        | 'VENUE_SELECTED'
        | 'REJECTED'
        | 'BOOKED'
        | 'COMPLETED'
        | 'CANCELLED';
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['input'];
    };
    readonly Invoice: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly bookingId: CodecTypes['pg/int4@1']['input'];
      readonly pdfUrl: CodecTypes['pg/text@1']['input'];
      readonly gstin: CodecTypes['pg/text@1']['input'];
    };
    readonly Policy: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly companyId: CodecTypes['pg/int4@1']['input'];
      readonly maxBudget: CodecTypes['pg/int4@1']['input'];
      readonly perPersonCap: CodecTypes['pg/int4@1']['input'];
      readonly allowedCities: ReadonlyArray<CodecTypes['pg/text@1']['input']>;
      readonly approverRole: 'ADMIN' | 'REQUESTER' | 'APPROVER' | 'FINANCE';
    };
    readonly User: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly clerkId: CodecTypes['pg/text@1']['input'];
      readonly email: CodecTypes['pg/text@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
      readonly role: 'ADMIN' | 'REQUESTER' | 'APPROVER' | 'FINANCE';
      readonly companyId: CodecTypes['pg/int4@1']['input'] | null;
    };
    readonly Venue: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly companyId: CodecTypes['pg/int4@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
      readonly city: CodecTypes['pg/text@1']['input'];
      readonly address: CodecTypes['pg/text@1']['input'] | null;
      readonly capacity: CodecTypes['pg/int4@1']['input'];
      readonly cuisine: CodecTypes['pg/text@1']['input'];
      readonly priceBand: CodecTypes['pg/text@1']['input'];
      readonly tags: ReadonlyArray<CodecTypes['pg/text@1']['input']>;
      readonly rating: CodecTypes['pg/float8@1']['input'];
      readonly active: CodecTypes['pg/bool@1']['input'];
    };
  };
};
// ...
export type TypeMaps = TypeMapsType<
  CodecTypes,
  QueryOperationTypes,
  FieldOutputTypes,
  FieldInputTypes,
  StorageColumnTypes,
  StorageColumnInputTypes,
  AggregateTypes
>;
// ...
export type Contract = ContractWithTypeMaps<ContractBase, TypeMaps>;